+++ react/AvailabilitySummary.tsx
@@ -28,17 +28,20 @@ const CSS_HANDLES = [
   'regularShipping',
   'time',
   'ETA',
+  'distance',
+  'distanceEstimate',
 ] as const
 
 interface CheckAvailabilityProps {
   maxItems: number
   orderBy: string
   pickupFirst: boolean
+  showDistance?: string
 }
 
 const AvailabilitySummary: StorefrontFunctionComponent<
   WrappedComponentProps & CheckAvailabilityProps
-> = ({ intl, maxItems, orderBy, pickupFirst }: any) => {
+> = ({ intl, maxItems, orderBy, pickupFirst, showDistance }: any) => {
   const [getSimulation, { data, loading }] = useLazyQuery(SIMULATE)
   const { data: orderFormData, refetch } = useQuery(ORDERFORM, { ssr: false })
 
@@ -71,6 +74,7 @@ const AvailabilitySummary: StorefrontFunctionComponent<
     const [logistics] = item.logisticsInfo
     const { availability } = status
     const { slas } = logistics
+    const kmToMile = 0.621371
 
     if (availability === 'withoutStock') return
     if (availability === 'cannotBeDelivered')
@@ -94,6 +98,8 @@ const AvailabilitySummary: StorefrontFunctionComponent<
           isPickup: !!option.pickupStoreInfo.address,
           storeName: option.pickupStoreInfo.friendlyName,
           days: parseInt(option.shippingEstimate.replace(/\D/g, ''), 10),
+          distance: option.pickupDistance,
+          showDistance,
         }
       })
       .sort((a: any, b: any) => {
@@ -211,6 +217,28 @@ const AvailabilitySummary: StorefrontFunctionComponent<
                   : `${styles.getInDays}`
               }`}
             >
+              {option.showDistance ? (
+                option.distance != null ? (
+                  <span className={styles.distance}>
+                    <FormattedMessage id="store/location-availability.distance.title" />{' '}
+                    <span className={styles.distanceEstimate}>
+                      {option.mesurements === 'kilometers'
+                        ? option.distance
+                        : option.distance * kmToMile}{' '}
+                    </span>
+                    {option.showDistance === 'miles' ? (
+                      <FormattedMessage id="store/location-availability.distance.miles" />
+                    ) : (
+                      <FormattedMessage id="store/location-availability.distance.kilometers" />
+                    )}{' '}
+                    <FormattedMessage id="store/location-availability.distance.away" />
+                  </span>
+                ) : (
+                  ''
+                )
+              ) : (
+                ''
+              )}
               <span className={styles.regularShippingLabel}>
                 <FormattedMessage id="store/location-availability.shipping-label" />
               </span>{' '}
@@ -315,6 +343,21 @@ const messages = defineMessages({
     id:
       'admin/editor.location-availability.product-location-availability.orderBy.cheaper',
   },
+  distance: {
+    defaultMessage: '',
+    id:
+      'admin/editor.location-availability.product-location-availability.distance.title',
+  },
+  miles: {
+    defaultMessage: '',
+    id:
+      'admin/editor.location-availability.product-location-availability.distance.miles',
+  },
+  kilometers: {
+    defaultMessage: '',
+    id:
+      'admin/editor.location-availability.product-location-availability.distance.kilometers',
+  },
 })
 
 AvailabilitySummary.schema = {
@@ -340,6 +383,14 @@ AvailabilitySummary.schema = {
       title: messages.pickupFirst,
       type: 'boolean',
       default: true,
+      isLayout: true,
+    },
+    distanceMeasurements: {
+      title: messages.distance,
+      type: 'string',
+      enum: ['miles', 'kilometers'],
+      enumNames: [messages.miles, messages.kilometers],
+      default: 'miles',
       isLayout: true,
     },
   },